LayoutTests/platform/chromium/editing/pasteboard/text-precomposed-expected.txt

Test that text retrieved from the clipboard is always precomposed.

PASS: text/plain data on the clipboard is precomposed
PASS: pasted text is precomposed
